requirements.txt
alembic==1.6.5
click==8.0.1
Flask==2.0.1
Flask-Migrate==3.0.1
Flask-SQLAlchemy==2.5.1
flask-swagger-ui==3.36.0
greenlet==1.1.0
itsdangerous==2.0.1
Jinja2==3.0.1
Mako==1.1.4
MarkupSafe==2.0.1
psycopg2-binary==2.9.1
PyJWT==2.1.0
python-dateutil==2.8.1
python-dotenv==0.18.0
python-editor==1.0.4
redis==3.5.3
six==1.16.0
SQLAlchemy==1.4.20
Werkzeug==2.0.1
